+++ src/main/java/insane96mcp/iguanatweaksexpanded/module/sleeprespawn/respawn/RespawnObeliskFeature.java
@@ -88,8 +88,13 @@ public class RespawnObeliskFeature extends JsonFeature {
 				|| !(event.getEntity() instanceof ServerPlayer player))
 			return;
 
-		if (player.getRespawnPosition() != null && player.level().getBlockState(player.getRespawnPosition()).is(RESPAWN_OBELISK.block().get()) && player.level().getBlockState(player.getRespawnPosition()).getValue(RespawnObeliskBlock.ENABLED)
-			&& event.getNewSpawn() != null && !player.level().getBlockState(event.getNewSpawn()).is(RESPAWN_OBELISK.block().get())) {
+		if (player.getRespawnPosition() != null
+				&& player.level().getBlockState(player.getRespawnPosition()).is(RESPAWN_OBELISK.block().get())
+				&& player.level().getBlockState(player.getRespawnPosition()).getValue(RespawnObeliskBlock.ENABLED)
+				&& event.getNewSpawn() != null
+				&& !player.level().getBlockState(event.getNewSpawn()).is(RESPAWN_OBELISK.block().get())) {
+			RespawnObeliskBlock.trySaveOldSpawn(player);
+			player.sendSystemMessage(Component.translatable("iguanatweaksexpanded.set_old_respawn"));
 			event.setCanceled(true);
 		}
 	}