+++ changelog.md
@@ -3,6 +3,7 @@
 ## Upcoming
 * Expanded is now compatible with Pickaxes and shovels
 * Respawn obelisks now have a slightly lower chance to have iron blocks
+* You can now change your old respawn point when you have a respawn obelisk set
 
 ## 1.1.3
 * Reduced mobs attack reach by ~15%